Rename TableroZoologicView to ZoologicBoardView
